fix: add tests for data adapter, fix lints
