+++ packages/amplify-gen1-codegen-data-adapter/src/index.ts
@@ -1,12 +1,1 @@
-import assert from 'node:assert';
-import { DataDefinition } from '@aws-amplify/amplify-gen2-codegen';
-import { Stack } from '@aws-sdk/client-cloudformation';
-
-const tableMappingKey = 'TableMapping';
-
-export const getDataDefinition = (dataStack: Stack): DataDefinition => {
-  const rawTableMapping = dataStack.Outputs?.find((o) => o.OutputKey === tableMappingKey)?.OutputValue;
-  assert(rawTableMapping);
-  const tableMapping = JSON.parse(rawTableMapping);
-  return { tableMapping };
-};
+export { getDataDefinition } from './get_data_definition';
